Split up the main file even more
- move functions into more new modules
- also add few more docstrings
- update tests to reflect all of the recent refactoring
